backend/requirements.txt
# Education Tutor Backend Requirements

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Embeddings and Vector Store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# LLM Integrations (optional)
openai>=1.3.0
google-generativeai>=0.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
